+++ tests/exporters/test_retry_fewer_spans.py
@@ -3,7 +3,6 @@ from __future__ import annotations
 from typing import Sequence
 
 import pytest
-from inline_snapshot import snapshot
 from opentelemetry.sdk.resources import Resource
 from opentelemetry.sdk.trace import Event, ReadableSpan
 from opentelemetry.sdk.trace.export import SpanExportResult
@@ -14,13 +13,10 @@ from opentelemetry.trace import SpanContext, SpanKind
 from opentelemetry.trace.status import Status, StatusCode
 
 from logfire._internal.exporters.otlp import (
-    BodySizeCheckingOTLPSpanExporter,
     BodyTooLargeError,
-    OTLPExporterHttpSession,
     RetryFewerSpansSpanExporter,
 )
 from logfire.testing import TestExporter
-from tests.exporters.test_otlp_session import SinkHTTPAdapter
 
 RESOURCE = Resource.create({'service.name': 'test', 'telemetry.sdk.version': '1.0.0'})
 TEST_SPANS = [
@@ -103,16 +99,3 @@ def test_retry_fewer_spans_when_too_many():
     res = exporter.export(TEST_SPANS)
     assert res is SpanExportResult.SUCCESS
     assert test_exporter.exported_spans == TEST_SPANS
-
-
-def test_max_body_size_bytes() -> None:
-    session = OTLPExporterHttpSession()
-    session.mount('http://', SinkHTTPAdapter())
-    exporter = BodySizeCheckingOTLPSpanExporter(session=session)
-
-    assert exporter.export(TEST_SPANS) == SpanExportResult.SUCCESS
-
-    exporter.max_body_size = 10
-    with pytest.raises(BodyTooLargeError) as e:
-        exporter.export(TEST_SPANS)
-    assert str(e.value) == snapshot('Request body is too large (897045 bytes), must be less than 10 bytes.')
